Restructure of files.
