+++ turtlemd/potentials/potential.py
@@ -12,7 +12,7 @@ from typing import TYPE_CHECKING, Any
 import numpy as np
 
 if TYPE_CHECKING:
-    from turtlemd.system import System  # pragma: no cover
+    from turtlemd.system.system import System  # pragma: no cover
 
 LOGGER = logging.getLogger(__name__)
 LOGGER.addHandler(logging.NullHandler())
