+++ turtlemd/system/particles.py
